Move custom tables controller to correct directory
