+++ tests/_support/Traits/Order_Modifiers.php
@@ -3,7 +3,7 @@
 namespace Tribe\Tickets\Test\Traits;
 
 use TEC\Common\StellarWP\Models\Contracts\Model;
-use TEC\Tickets\Commerce\Order_Modifiers\Controller;
+use TEC\Tickets\Commerce\Order_Modifiers\Custom_Tables\Controller;
 use TEC\Tickets\Commerce\Order_Modifiers\Models\Order_Modifier;
 use TEC\Tickets\Commerce\Order_Modifiers\Modifiers\Modifier_Manager;
 use TEC\Tickets\Commerce\Order_Modifiers\Repositories\Order_Modifiers as Order_Modifiers_Repository;
